GUI lfo + transpose fix

+++ src/oscillator.cpp
@@ -5,7 +5,7 @@
 
 #define M_PI 3.14159265358979323846
 
-Osc::Osc(double v, Waves w, double lfo_h, double lfo_a, int t)
+Osc::Osc(double v, Waves w, double lfo_h, double lfo_a, double t)
     : volume(v),
       wave(w),
       lfo_hertz(lfo_h),
@@ -14,7 +14,7 @@ Osc::Osc(double v, Waves w, double lfo_h, double lfo_a, int t)
 
 double Osc::sound(double time, double frequency) {
     
-    frequency = frequency * pow(2, transpose / 12);
+    frequency = frequency * pow(2, transpose / 12.0);
 
     double amp = w(frequency) * time + lfo_amplitude * frequency * sin(w(lfo_hertz) * time);
     double output;
